lib: separate maildir flags per file
We do not store maildir flags in the database, only tags which derive
from them. To do that, message->maildir_flags keeps a concatenation of
all maildir info strings from all files belonging to that message. So,
be testing that string for the presence of a character A, one tests for
the presence of that maildir flag on at least one file, effectly
implementing an OR on the flags.

Change this slightly so that message->maildir_flags separates individual
info strings by a comma. This does not alter any flag logic, costs 1
byte per message structure that we hold in memory, but allows to do
several things later on:

- count the info strings easily (without walking a list)
- count the occurrences of a flag and compare it to the number above
- check for the presence of specific combinations of flags om individual
  message files

This will be used (at least) for implementing an AND on flags where this
gives a better user experience than OR.

Note that the AND that we effectively have for "unread" already simply
comes from the fact that
  unread [the tag] == NOT seen [the maildir flag S]
and
  NOT (S1 OR S2) == (NOT S1) AND (NOT S2).

An alternative approach would store the flags in the database per message
file, deviating from the "one mid one message" concept, but still facing
the problem of how to map this to message tags.