lib: Reinstate synchronization of 'T' flag with "deleted" tag.
2c262042 ("lib: Remove the synchronization of 'T' flag with "deleted"
tag.", 2010-11-09) justified the removal with the case of multiple files
for one message with different flags, which in some workflows does not
indicate an intention to remove the message (but only one copy).

Therefore, reinstate the synchronization as an AND of the individual
flags: a message is tagged "deleted" only if all its associated files
carry the "T" flag.

This avoids accidental deletions in the mentioned workflow (including
the dump/restore pitfall cited in 2c262042).

Conversely, tagging a message "deleted" clearly indicates the intention
to delete it, and "T" is set on all file copies.

Note that we stick to the tag name "deleted" (rather than "trashed") for
"T" for legacy reasons only.